+++ web/src/add_subscribers_pill.ts
@@ -55,6 +55,7 @@ function set_up_pill_typeahead({
         stream: true,
         user_group: true,
         user: true,
+        only_show_user_groups_editable_by_user: false,
     };
     pill_typeahead.set_up_combined($pill_container.find(".input"), pill_widget, opts);
 }