+++ zerver/actions/invites.py
@@ -31,7 +31,15 @@ from zerver.lib.queue import queue_event_on_commit
 from zerver.lib.send_email import FromAddress, clear_scheduled_invitation_emails, send_future_email
 from zerver.lib.timestamp import datetime_to_timestamp
 from zerver.lib.utils import assert_is_not_none
-from zerver.models import Message, MultiuseInvite, PreregistrationUser, Realm, Stream, UserProfile
+from zerver.models import (
+    Message,
+    MultiuseInvite,
+    NamedUserGroup,
+    PreregistrationUser,
+    Realm,
+    Stream,
+    UserProfile,
+)
 from zerver.models.prereg_users import filter_to_valid_prereg_users
 
 
@@ -174,6 +182,7 @@ def do_invite_users(
     user_profile: UserProfile,
     invitee_emails: Collection[str],
     streams: Collection[Stream],
+    user_groups: Collection[NamedUserGroup] = [],
     *,
     invite_expires_in_minutes: Optional[int],
     include_realm_default_subscriptions: bool,
@@ -268,6 +277,8 @@ def do_invite_users(
         prereg_user.save()
         stream_ids = [stream.id for stream in streams]
         prereg_user.streams.set(stream_ids)
+        user_groups_ids = [user_group.id for user_group in user_groups]
+        prereg_user.user_groups.set(user_groups_ids)
 
         confirmation = create_confirmation_object(
             prereg_user, Confirmation.INVITATION, validity_in_minutes=invite_expires_in_minutes
@@ -362,6 +373,7 @@ def do_create_multiuse_invite_link(
     invite_expires_in_minutes: Optional[int],
     include_realm_default_subscriptions: bool,
     streams: Sequence[Stream] = [],
+    user_groups: Sequence[NamedUserGroup] = [],
 ) -> str:
     realm = referred_by.realm
     invite = MultiuseInvite.objects.create(
@@ -371,6 +383,8 @@ def do_create_multiuse_invite_link(
     )
     if streams:
         invite.streams.set(streams)
+    if user_groups:
+        invite.user_groups.set(user_groups)
     invite.invited_as = invited_as
     invite.save()
     notify_invites_changed(referred_by.realm, changed_invite_referrer=referred_by)