+++ zerver/models/prereg_users.py
@@ -70,6 +70,7 @@ class PreregistrationUser(models.Model):
     full_name_validated = models.BooleanField(default=False)
     referred_by = models.ForeignKey(UserProfile, null=True, on_delete=CASCADE)
     streams = models.ManyToManyField("zerver.Stream")
+    user_groups = models.ManyToManyField("zerver.NamedUserGroup")
     invited_at = models.DateTimeField(auto_now=True)
     realm_creation = models.BooleanField(default=False)
     # Indicates whether the user needs a password.  Users who were
@@ -130,6 +131,7 @@ def filter_to_valid_prereg_users(
 class MultiuseInvite(models.Model):
     referred_by = models.ForeignKey(UserProfile, on_delete=CASCADE)
     streams = models.ManyToManyField("zerver.Stream")
+    user_groups = models.ManyToManyField("zerver.NamedUserGroup")
     realm = models.ForeignKey(Realm, on_delete=CASCADE)
     invited_as = models.PositiveSmallIntegerField(default=PreregistrationUser.INVITE_AS["MEMBER"])
     include_realm_default_subscriptions = models.BooleanField(default=True)