+++ zerver/tests/test_invite.py
@@ -27,7 +27,7 @@ from zerver.actions.create_realm import do_change_realm_subdomain, do_create_rea
 from zerver.actions.create_user import (
     do_create_user,
     process_new_human_user,
-    set_up_streams_for_new_human_user,
+    set_up_streams_and_groups_for_new_human_user,
 )
 from zerver.actions.default_streams import do_add_default_stream, do_remove_default_stream
 from zerver.actions.invites import (
@@ -42,6 +42,7 @@ from zerver.actions.realm_settings import (
     do_change_realm_plan_type,
     do_set_realm_property,
 )
+from zerver.actions.user_groups import create_user_group_in_database
 from zerver.actions.user_settings import do_change_full_name
 from zerver.actions.users import change_user_is_active
 from zerver.context_processors import common_context
@@ -54,6 +55,7 @@ from zerver.lib.send_email import FromAddress, deliver_scheduled_emails, send_fu
 from zerver.lib.streams import ensure_stream
 from zerver.lib.test_classes import ZulipTestCase
 from zerver.lib.test_helpers import find_key_by_email
+from zerver.lib.user_groups import get_direct_user_groups
 from zerver.models import (
     DefaultStream,
     Message,
@@ -110,7 +112,7 @@ class StreamSetupTest(ZulipTestCase):
         new_user = self.create_simple_new_user(realm, "alice@zulip.com")
 
         with self.assert_database_query_count(13):
-            set_up_streams_for_new_human_user(
+            set_up_streams_and_groups_for_new_human_user(
                 user_profile=new_user,
                 prereg_user=None,
                 default_stream_groups=[],
@@ -145,8 +147,40 @@ class StreamSetupTest(ZulipTestCase):
 
         new_user = self.create_simple_new_user(realm, new_user_email)
 
-        with self.assert_database_query_count(13):
-            set_up_streams_for_new_human_user(
+        with self.assert_database_query_count(14):
+            set_up_streams_and_groups_for_new_human_user(
+                user_profile=new_user,
+                prereg_user=prereg_user,
+                default_stream_groups=[],
+            )
+
+    def test_query_count_when_admin_assigns_groups(self) -> None:
+        admin = self.example_user("iago")
+        realm = admin.realm
+
+        hamletcharacters_group = NamedUserGroup.objects.get(name="hamletcharacters", realm=realm)
+        test_group = create_user_group_in_database("test", [admin], realm, acting_user=None)
+        user_groups = [hamletcharacters_group, test_group]
+
+        self.add_messages_to_stream("Rome")
+
+        new_user_email = "bob@zulip.com"
+
+        do_invite_users(
+            admin,
+            [new_user_email],
+            streams=[],
+            user_groups=user_groups,
+            include_realm_default_subscriptions=False,
+            invite_expires_in_minutes=1000,
+        )
+
+        prereg_user = PreregistrationUser.objects.get(email=new_user_email)
+
+        new_user = self.create_simple_new_user(realm, new_user_email)
+
+        with self.assert_database_query_count(9):
+            set_up_streams_and_groups_for_new_human_user(
                 user_profile=new_user,
                 prereg_user=prereg_user,
                 default_stream_groups=[],
@@ -175,6 +209,7 @@ class InviteUserBase(ZulipTestCase):
         self,
         invitee_emails: str,
         stream_names: Sequence[str],
+        user_group_ids: Sequence[int] = [],
         invite_expires_in_minutes: Optional[int] = INVITATION_LINK_VALIDITY_MINUTES,
         body: str = "",
         invite_as: int = PreregistrationUser.INVITE_AS["MEMBER"],
@@ -188,6 +223,8 @@ class InviteUserBase(ZulipTestCase):
             newline separated.
 
         streams should be a list of strings.
+
+        user_group_ids should be a list of user_groups ids.
         """
         stream_ids = [self.get_stream_id(stream_name, realm=realm) for stream_name in stream_names]
 
@@ -202,6 +239,7 @@ class InviteUserBase(ZulipTestCase):
                     "invitee_emails": invitee_emails,
                     "invite_expires_in_minutes": invite_expires_in,
                     "stream_ids": orjson.dumps(stream_ids).decode(),
+                    "user_group_ids": orjson.dumps(user_group_ids).decode(),
                     "invite_as": invite_as,
                     "include_realm_default_subscriptions": orjson.dumps(
                         include_realm_default_subscriptions
@@ -730,6 +768,23 @@ class InviteUserTest(InviteUserBase):
         response = self.invite(invitee, ["Denmark"], invite_as=10)
         self.assert_json_error(response, "Invalid invite_as")
 
+    def test_invite_user_with_specified_user_groups_when_cannot_edit_user_groups(self) -> None:
+        hamlet = self.example_user("hamlet")
+        realm = hamlet.realm
+        user_group = create_user_group_in_database("test", [hamlet], realm, acting_user=None)
+        # Only admins are allowed to create and manage user groups.
+        do_set_realm_property(
+            realm,
+            "user_group_edit_policy",
+            Realm.POLICY_ADMINS_ONLY,
+            acting_user=None,
+        )
+
+        self.login("hamlet")
+        invitee = self.nonreg_email("bob")
+        result = self.invite(invitee, [], [user_group.id])
+        self.assert_json_error(result, "Insufficient permission")
+
     def test_successful_invite_user_as_guest_from_normal_account(self) -> None:
         self.login("hamlet")
         invitee = self.nonreg_email("alice")
@@ -838,6 +893,29 @@ class InviteUserTest(InviteUserBase):
         # passed in streams list.
         self.submit_reg_form_for_user(invitee, "password")
         self.check_user_subscribed_only_to_streams("test1", [denmark, sandbox, verona, zulip])
+
+    def test_successful_invite_users_with_specified_user_groups(self) -> None:
+        invitee = self.nonreg_email("bob")
+        iago = self.example_user("iago")
+        self.login("iago")
+
+        user_group1 = create_user_group_in_database("test1", [], iago.realm, acting_user=iago)
+        user_group2 = create_user_group_in_database("test2", [], iago.realm, acting_user=iago)
+
+        user_group_ids = [user_group1.id, user_group2.id]
+
+        self.assert_json_success(self.invite(invitee, [], user_group_ids))
+        self.assertTrue(find_key_by_email(invitee))
+        self.submit_reg_form_for_user(invitee, "password")
+
+        # bob is a direct member of two role-based system groups also.
+        user_groups_subscriptions = get_direct_user_groups(self.nonreg_user("bob"))
+        user_group_names = [group.named_user_group.name for group in user_groups_subscriptions]
+
+        self.assertEqual(
+            set(user_group_names),
+            {"test1", "test2", SystemGroups.MEMBERS, SystemGroups.FULL_MEMBERS},
+        )
 
     def test_can_invite_others_to_realm(self) -> None:
         def validation_func(user_profile: UserProfile) -> bool:
@@ -1094,6 +1172,32 @@ earl-test@zulip.com""",
         self.assert_json_error(
             self.invite("iago-test@zulip.com", ["NotARealStream"]),
             f"Invalid channel ID {self.INVALID_STREAM_ID}. No invites were sent.",
+        )
+        self.check_sent_emails([])
+
+    def test_invalid_user_group(self) -> None:
+        """
+        Tests inviting to a non-existent or inaccessible UserGroup.
+        """
+        self.login("hamlet")
+        self.assert_json_error(
+            self.invite("iago-test@zulip.com", ["Denmark"], [5678]),
+            "Usergroup does not exist with id: 5678. No invites were sent.",
+        )
+        self.check_sent_emails([])
+
+        hamlet = self.example_user("hamlet")
+        realm = hamlet.realm
+        self.login("hamlet")
+
+        user_group = create_user_group_in_database(
+            "test", [self.example_user("cordelia")], realm, acting_user=None
+        )
+
+        self.assertEqual(realm.user_group_edit_policy, Realm.POLICY_MEMBERS_ONLY)
+        self.assert_json_error(
+            self.invite("iago-test@zulip.com", ["Denmark"], [user_group.id]),
+            f"Usergroup does not exist with id: {user_group.id}. No invites were sent.",
         )
         self.check_sent_emails([])
 
@@ -2302,7 +2406,7 @@ class InvitationsTestCase(InviteUserBase):
         self.login("iago")
         invitee = "resend@zulip.com"
 
-        self.assert_json_success(self.invite(invitee, ["Denmark"], None))
+        self.assert_json_success(self.invite(invitee, ["Denmark"], [], None))
         prereg_user = PreregistrationUser.objects.get(email=invitee)
 
         # Verify and then clear from the outbox the original invite email
@@ -2402,6 +2506,7 @@ class MultiuseInviteTest(ZulipTestCase):
         streams: Optional[List[Stream]] = None,
         date_sent: Optional[datetime] = None,
         include_realm_default_subscriptions: bool = False,
+        user_groups: Optional[List[NamedUserGroup]] = None,
     ) -> str:
         invite = MultiuseInvite(
             realm=self.realm,
@@ -2412,6 +2517,9 @@ class MultiuseInviteTest(ZulipTestCase):
 
         if streams is not None:
             invite.streams.set(streams)
+
+        if user_groups is not None:
+            invite.user_groups.set(user_groups)
 
         if date_sent is None:
             date_sent = timezone_now()
@@ -2551,6 +2659,27 @@ class MultiuseInviteTest(ZulipTestCase):
         rome = get_stream("Rome", self.realm)
         self.check_user_subscribed_only_to_streams(
             name5, [rome, default_streams[0], default_streams[1], default_streams[2]]
+        )
+
+    def test_multiuse_link_with_specified_user_groups(self) -> None:
+        iago = self.example_user("iago")
+        self.login("iago")
+
+        user_group1 = create_user_group_in_database("test1", [], iago.realm, acting_user=iago)
+        user_group2 = create_user_group_in_database("test2", [], iago.realm, acting_user=iago)
+
+        user_groups = [user_group1, user_group2]
+
+        invite_link = self.generate_multiuse_invite_link(user_groups=user_groups)
+        self.check_user_able_to_register(self.nonreg_email("bob"), invite_link)
+
+        # bob is a direct member of two role-based system groups also.
+        user_groups_subscriptions = get_direct_user_groups(self.nonreg_user("bob"))
+        user_group_names = [group.named_user_group.name for group in user_groups_subscriptions]
+
+        self.assertEqual(
+            set(user_group_names),
+            {"test1", "test2", SystemGroups.MEMBERS, SystemGroups.FULL_MEMBERS},
         )
 
     def test_multiuse_link_different_realms(self) -> None:
@@ -2729,6 +2858,61 @@ class MultiuseInviteTest(ZulipTestCase):
         )
         self.assert_json_success(result)
 
+    def test_multiuser_link_with_specified_user_groups_when_cannot_edit_user_groups(self) -> None:
+        hamlet = self.example_user("hamlet")
+        realm = hamlet.realm
+        user_group = create_user_group_in_database("test", [hamlet], realm, acting_user=None)
+        # Only admins are allowed to create and manage user groups.
+        do_set_realm_property(
+            realm,
+            "user_group_edit_policy",
+            Realm.POLICY_ADMINS_ONLY,
+            acting_user=None,
+        )
+
+        members_system_group = NamedUserGroup.objects.get(
+            name=SystemGroups.MEMBERS, realm=realm, is_system_group=True
+        )
+        do_change_realm_permission_group_setting(
+            realm, "create_multiuse_invite_group", members_system_group, acting_user=None
+        )
+        self.login("hamlet")
+        result = self.client_post(
+            "/json/invites/multiuse",
+            {
+                "user_group_ids": orjson.dumps([user_group.id]).decode(),
+                "invite_expires_in_minutes": 2 * 24 * 60,
+            },
+        )
+        self.assert_json_error(result, "Insufficient permission")
+
+    def test_create_multiuse_link_with_specified_user_groups_api_call(self) -> None:
+        iago = self.example_user("iago")
+        self.login("iago")
+
+        user_group1 = create_user_group_in_database("test1", [], iago.realm, acting_user=iago)
+        user_group2 = create_user_group_in_database("test2", [], iago.realm, acting_user=iago)
+
+        user_group_ids = [user_group1.id, user_group2.id]
+        result = self.client_post(
+            "/json/invites/multiuse",
+            {
+                "user_group_ids": orjson.dumps(user_group_ids).decode(),
+                "invite_expires_in_minutes": 2 * 24 * 60,
+            },
+        )
+        invite_link = self.assert_json_success(result)["invite_link"]
+        self.check_user_able_to_register(self.nonreg_email("bob"), invite_link)
+
+        # bob is a direct member of two role-based system groups also.
+        user_groups_subscriptions = get_direct_user_groups(self.nonreg_user("bob"))
+        user_group_names = [group.named_user_group.name for group in user_groups_subscriptions]
+
+        self.assertEqual(
+            set(user_group_names),
+            {"test1", "test2", SystemGroups.MEMBERS, SystemGroups.FULL_MEMBERS},
+        )
+
     def test_create_multiuse_invite_group_setting(self) -> None:
         realm = get_realm("zulip")
         full_members_system_group = NamedUserGroup.objects.get(
@@ -2890,6 +3074,42 @@ class MultiuseInviteTest(ZulipTestCase):
             },
         )
         self.assert_json_error(result, "Invalid channel ID 54321. No invites were sent.")
+
+    def test_create_multiuse_link_invalid_user_group_api_call(self) -> None:
+        self.login("iago")
+        result = self.client_post(
+            "/json/invites/multiuse",
+            {
+                "user_group_ids": orjson.dumps([5438]).decode(),
+                "invite_expires_in_minutes": 2 * 24 * 60,
+            },
+        )
+        self.assert_json_error(result, "Invalid user_group ID 5438. No invites were sent.")
+
+        # Test the case when user cannot access the user group.
+        realm = get_realm("zulip")
+        members_system_group = NamedUserGroup.objects.get(name=SystemGroups.MEMBERS, realm=realm)
+        do_change_realm_permission_group_setting(
+            realm, "create_multiuse_invite_group", members_system_group, acting_user=None
+        )
+
+        self.login("hamlet")
+        user_group = create_user_group_in_database(
+            "test", [self.example_user("cordelia")], get_realm("zulip"), acting_user=None
+        )
+
+        self.assertEqual(realm.user_group_edit_policy, Realm.POLICY_MEMBERS_ONLY)
+
+        result = self.client_post(
+            "/json/invites/multiuse",
+            {
+                "user_group_ids": orjson.dumps([user_group.id]).decode(),
+                "invite_expires_in_minutes": 2 * 24 * 60,
+            },
+        )
+        self.assert_json_error(
+            result, f"Invalid user_group ID {user_group.id}. No invites were sent."
+        )
 
     def test_create_multiuse_link_invalid_invite_as_api_call(self) -> None:
         self.login("iago")