+++ zerver/views/invite.py
@@ -2,6 +2,7 @@ import re
 from typing import List, Optional, Sequence, Set
 
 from django.conf import settings
+from django.db import transaction
 from django.http import HttpRequest, HttpResponse
 from django.utils.timezone import now as timezone_now
 from django.utils.translation import gettext as _
@@ -20,8 +21,9 @@ from zerver.lib.exceptions import InvitationError, JsonableError, OrganizationOw
 from zerver.lib.request import REQ, has_request_variables
 from zerver.lib.response import json_success
 from zerver.lib.streams import access_stream_by_id
+from zerver.lib.user_groups import access_user_group_by_id
 from zerver.lib.validator import check_bool, check_int, check_int_in, check_list, check_none_or
-from zerver.models import MultiuseInvite, PreregistrationUser, Stream, UserProfile
+from zerver.models import MultiuseInvite, NamedUserGroup, PreregistrationUser, Stream, UserProfile
 
 # Convert INVITATION_LINK_VALIDITY_DAYS into minutes.
 # Because mypy fails to correctly infer the type of the validator, we want this constant
@@ -60,6 +62,7 @@ def invite_users_backend(
     ),
     stream_ids: List[int] = REQ(json_validator=check_list(check_int)),
     include_realm_default_subscriptions: bool = REQ(json_validator=check_bool, default=False),
+    user_group_ids: Sequence[int] = REQ(json_validator=check_list(check_int), default=[]),
 ) -> HttpResponse:
     if not user_profile.can_invite_users_by_email():
         # Guest users case will not be handled here as it will
@@ -95,10 +98,27 @@ def invite_users_backend(
     if len(streams) and not user_profile.can_subscribe_other_users():
         raise JsonableError(_("You do not have permission to subscribe other users to channels."))
 
+    user_groups: List[NamedUserGroup] = []
+    with transaction.atomic():
+        for user_group_id in user_group_ids:
+            try:
+                user_group = access_user_group_by_id(user_group_id, user_profile, for_read=False)
+            except JsonableError:
+                raise JsonableError(
+                    _(
+                        "Usergroup does not exist with id: {user_group_id}. No invites were sent."
+                    ).format(user_group_id=user_group_id)
+                )
+            user_groups.append(user_group)
+
+    if user_groups and not user_profile.can_edit_user_groups():
+        raise JsonableError(_("Insufficient permission"))
+
     skipped = do_invite_users(
         user_profile,
         invitee_emails,
         streams,
+        user_groups,
         invite_expires_in_minutes=invite_expires_in_minutes,
         include_realm_default_subscriptions=include_realm_default_subscriptions,
         invite_as=invite_as,
@@ -216,6 +236,7 @@ def generate_multiuse_invite_backend(
     ),
     stream_ids: Sequence[int] = REQ(json_validator=check_list(check_int), default=[]),
     include_realm_default_subscriptions: bool = REQ(json_validator=check_bool, default=False),
+    user_group_ids: Sequence[int] = REQ(json_validator=check_list(check_int), default=[]),
 ) -> HttpResponse:
     if not user_profile.can_create_multiuse_invite_to_realm():
         # Guest users case will not be handled here as it will
@@ -246,11 +267,28 @@ def generate_multiuse_invite_backend(
     if len(streams) and not user_profile.can_subscribe_other_users():
         raise JsonableError(_("You do not have permission to subscribe other users to channels."))
 
+    user_groups: List[NamedUserGroup] = []
+    with transaction.atomic():
+        for user_group_id in user_group_ids:
+            try:
+                user_group = access_user_group_by_id(user_group_id, user_profile, for_read=False)
+            except JsonableError:
+                raise JsonableError(
+                    _("Invalid user_group ID {user_group_id}. No invites were sent.").format(
+                        user_group_id=user_group_id
+                    )
+                )
+            user_groups.append(user_group)
+
+    if user_groups and not user_profile.can_edit_user_groups():
+        raise JsonableError(_("Insufficient permission"))
+
     invite_link = do_create_multiuse_invite_link(
         user_profile,
         invite_as,
         invite_expires_in_minutes,
         include_realm_default_subscriptions,
         streams,
+        user_groups,
     )
     return json_success(request, data={"invite_link": invite_link})